2.碎片图像复原----第二版/Debug/requirements.txt
pillow
numpy
pyqt5